requirements.txt
mysql-connector-python
python-dotenv
langchain
openai
